Move testing code in svm_runtime into svm_runtime_testing
