+++ crates/runtime-ffi/Cargo.toml
@@ -18,6 +18,7 @@ log = "0.4.14"
 lazy_static = "1.4.0"
 svm-runtime = { path="../runtime", default-features=false }
 svm-gas = { path="../gas" }
+svm-genesis-config = { path="../genesis-config" }
 svm-types = { path="../types" }
 svm-codec = { path="../codec" }
 svm-layout = { path="../layout" }
@@ -28,8 +29,9 @@ wasmer = { version = "2.1.1", features = ["default-singlepass", "default-univers
 wat = "1.0.40"
 serde_json = "1.0.74"
 maplit = "1.0.2"
-svm-sdk-mock = { path="../sdk-mock", features = ["dynamic-alloc"], default-features = false }
 svm-abi-encoder = { path="../abi-encoder", features = ["dynamic-alloc"], default-features = false }
+svm-runtime-testing = { path = "../runtime-testing" }
+svm-sdk-mock = { path="../sdk-mock", features = ["dynamic-alloc"], default-features = false }
 svm-types = { path="../types", default-features = false }
 
 [build-dependencies]
