+++ crates/runtime-testing/src/lib.rs
@@ -1,14 +1,14 @@
 //! Implements common functionality to be consumed by tests.
 
 use svm_codec::{template, Codec};
+use svm_genesis_config::GenesisConfig;
 use svm_layout::{FixedLayout, Layout};
-use svm_state::{GenesisConfig, GlobalState};
+use svm_runtime::{PriceResolverRegistry, Runtime, TemplatePriceCache};
+use svm_state::GlobalState;
 use svm_types::{
     Address, CodeSection, CtorsSection, DataSection, HeaderSection, SpawnAccount, Template,
     TemplateAddr, Transaction,
 };
-
-use crate::{PriceResolverRegistry, Runtime, TemplatePriceCache};
 
 /// Hold a Wasm file in textual or binary form
 pub enum WasmFile<'a> {
