+++ crates/state/Cargo.toml
@@ -17,7 +17,7 @@ futures = "0.3.19"
 # https://github.com/launchbadge/sqlx/issues/713
 sqlx = { version = "0.5.10", features = ["runtime-tokio-rustls", "sqlite", "macros"] }
 svm-codec = { path = "../codec" }
-svm-genesis-templates = { path = "../genesis-templates" }
+svm-genesis-config = { path = "../genesis-config" }
 svm-hash = { path = "../hash" }
 svm-layout = { path = "../layout" }
 svm-types = { path = "../types" }
